Adds function overloading for vertex constructors and initialize missing containers with empty Eigen matrix.
Fixes a bug where observed_semantic_landmark_ids_ was not sized properly in the case of zero semantic landmarks.
We not add a check to resize the container to the same size as the number of frames to avoid access failure during run time.

Adds deserialization for observed_semantic_landmark_ids_